requirements.txt
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.26
langchain-experimental==0.0.52
langchain-openai==0.0.6
langsmith==0.1.5
matplotlib==3.8.0
numpy==1.26.3
numpydoc==1.5.0
openai==1.12.0
pydantic==1.10.12
opencv-python==4.9.0.80
ai2thor==5.0.0
weaviate-client
transformers==4.38.2
pillow==10.0.1
torch==2.2.1
